+++ crates/hyperqueue/src/server/autoalloc/queue/slurm.rs
@@ -57,8 +57,13 @@ impl QueueHandler for SlurmHandler {
                 allocation_num,
             )?;
 
-            let worker_args =
-                build_worker_args(&hq_path, ManagerType::Slurm, &server_directory, &params);
+            let worker_args = build_worker_args(
+                &hq_path,
+                ManagerType::Slurm,
+                &server_directory,
+                &params,
+                &working_dir,
+            )?;
             let worker_args = add_start_stop_worker_commands(
                 worker_args,
                 params.worker_start_cmd.as_deref(),